tests(e2e): refactor fixtures into core
